Add format and categories to assignment serializer
Due to the many M2M relationships introduced in this PR,
serializers are prone to circular serialization or duplicate
data.
Therefore I opted to simply serialize categories in the
respective assignment serializer. This avoids duplicate
and circualr serialization, the trick is to make the
assignment data always available frontend. And
preferably, to cache the assignment data, since
the call is quite expensive.

For example, Category serializes all (default) templates and
all available templates. These are required for a student
to see which defaults exists and which are selected by default.
Each template exists of a field set, which forms the corresponding
template's structure.
When serializing an entry, the corresponding template (deadline/edit)
is also serialized. Likewise all categories need to be available, which
in turn would need to serialize their corresponding templates.