Add mapping for filtered timeline nodes
Ugly solution to code which is long due for a refactor.

Mapping indices between all nodes and filtered nodes
meant the existing parent components which use the
TimeLine did not have to be touched.

However, because timeline nodes can consist of two
types of nodes (preset and default nodes), and a deep
emit chain exists for indicating which node is selected
(via index instead of object), I think this solution is my
most ugly/fragile commit of 2020.