Improved support for LTI test student accounts (#1132)
* Remove specific test user full_name check

A test user can be launched with different names, something we learned after observing a test user launch with the full_name 'Cusist toetsen' instead of the expected 'Test user'.

Whilst the params for a test user on first launch are LMS depedent, for Canvas we can move towards using the following solution:

If you are using the instructor student view, the LTI custom parameter Canvas.masqueradingUser.id will be set.  This alone won't let you know its the student view, but would let you LTI know its a masquerade.  There might be other LTI parameters that would help tie it down.  For example, lis_person_sourcedid seems to be missing for student view, but is there for an admin masquerading, so that might be a way of distinguishing the cases.

For now we have no normal users without an email adress, so loosening the check for lack of email value will suffice for now.

* More strictly log test student creation

* Add test student warning front end

* Throw exception on existing test student launch with email adress

* More human friendly test student with email launch error msg

* Fixed spelling

Co-authored-by: Lars <larsvanhijfte@hotmail.com>
Co-authored-by: Engel Hamer <39912581+engelhamer@users.noreply.github.com>