Fixed an issue where the timeline assignment could be undefined (#1140)
* Fixed an issue where the timeline assignment could be undefined

This fix is mostly a bandaid on top of the multiple calls to assignment retrieve with different arguments throughout the app.

A proper solution requires a refactor: isolate expensive assignment serializer logic to dedicated APIs. E.g. stats, journals, and courses.

* Actually add the conditional rendering

Bloody load wrapper not being a guard to the content itself

Co-authored-by: Lars <larsvanhijfte@hotmail.com>